refactor: separate files in pages/shared
