+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import './App.css'
-import Home from './components/Home'
+import Home from './components/pages/Home'
 
 function App() {
   return (
