+++ src/components/pages/Navbar.tsx
@@ -1,14 +1,14 @@
 import React from 'react'
-import { ReactComponent as LoginIcon } from '../icons/loginIcon.svg'
-import { ReactComponent as MenuIcon } from '../icons/menuIcon.svg'
-import { ReactComponent as SearchIcon } from '../icons/searchIcon.svg'
-import { ReactComponent as SignUpIcon } from '../icons/signupIcon.svg'
-import { ReactComponent as UserIcon } from '../icons/userIcon.svg'
-import { ReactComponent as MainLogo } from '../images/mainLogo.svg'
-import { ReactComponent as MobileLogo } from '../images/mobileLogo.svg'
-import Button from './Button'
-import DropdownContainer from './DropdownContainer'
-import Input from './Input'
+import { ReactComponent as LoginIcon } from '../../icons/loginIcon.svg'
+import { ReactComponent as MenuIcon } from '../../icons/menuIcon.svg'
+import { ReactComponent as SearchIcon } from '../../icons/searchIcon.svg'
+import { ReactComponent as SignUpIcon } from '../../icons/signupIcon.svg'
+import { ReactComponent as UserIcon } from '../../icons/userIcon.svg'
+import { ReactComponent as MainLogo } from '../../images/mainLogo.svg'
+import { ReactComponent as MobileLogo } from '../../images/mobileLogo.svg'
+import Button from '../shared/Button'
+import DropdownContainer from '../shared/DropdownContainer'
+import Input from '../shared/Input'
 import SidebarMobileContainer from './SidebarMobileContainer'
 
 export default function Navbar() {
